+++ src/solareclipseworkbench/scripts.py
@@ -40,6 +40,10 @@ def convert_command(line, ref_moment, sign, time_delta, extra_comment, output_fi
         comment = command_parts[12] if len(command_parts) > 12 else ""
         bracket = ""  # Will need to be set elsewhere
         command = "take_bracket"
+    elif command_parts[0] == "take_hdr":
+        camera_name, exposure, aperture, iso, stops = command_parts[4:9]
+        comment = command_parts[9] if len(command_parts) > 9 else ""
+        command = "take_hdr"
     elif command_parts[0] == "sync_cameras":
         comment = command_parts[4] if len(command_parts) > 4 else ""
         command = "sync_cameras"
@@ -82,6 +86,9 @@ def convert_command(line, ref_moment, sign, time_delta, extra_comment, output_fi
     elif command == "take_bracket":
         output_file.write(
             f"{command}, {ref_moment}, {sign}, {time_delta}, {camera_name.strip()}, {exposure.strip()}, {aperture.strip()}, {iso.strip()}, {bracket.strip()}, \"{comment.strip()+extra_comment}\"\n")
+    elif command == "take_hdr":
+        output_file.write(
+            f"{command}, {ref_moment}, {sign}, {time_delta}, {camera_name.strip()}, {exposure.strip()}, {aperture.strip()}, {iso.strip()}, {stops.strip()}, \"{comment.strip()+extra_comment}\"\n")
     elif command == "sync_cameras":
         output_file.write(f"{command}, {ref_moment}, {sign}, {time_delta}, \"{comment.strip()+extra_comment}\"\n")
     elif command == "voice_prompt":